Split out bounded streaming to new file
